refactor(training): 抽 anima_train.py 到 runtime/training/ 子包 (ADR 0003 PR-A) (#56)
* docs(adr): 0003 anima_train.py 模块化重构设计 + 修 0002 行号引用

ADR 0003 提议把 runtime/anima_train.py（2901 行 / main() 793 行）拆到
runtime/training/ 子包，分三个 PR 推进（A 搬运 / B phase 化 / C plugin + adapter
hook protocol）。含 9 个真实变体落地案例（DoRA / VeRA / T-LoRA / OFT /
Ortho-Hydra / Lion / warmup_cosine / Euler / 删除）。

同笔顺手修 0002 的两处 L2374-L2401 行号引用——重构后位置会变，改成 grep
关键词 (signal_handler / signal.SIGINT) + ADR 0003 重构后路径提示。


* refactor(anima_train): 抽 bootstrap / observability / model_loading / text_encoding / state 到 training/

ADR 0003 PR-A 第 1 步：把 anima_train.py 里 5 个无外部依赖的 util 簇按职责
搬到 runtime/training/ 子包：

- training/bootstrap.py        ← deps / yaml / 进度条 (5 fn)
- training/observability.py    ← WandBMonitor + loss 曲线 (4 entry)
- training/model_loading.py    ← prefix 推断 / safetensors / 路径解析 / xformers (9 fn)
- training/text_encoding.py    ← Qwen + T5 加权 tokenize (4 fn)
- training/state.py            ← save/load_training_state

顶层 anima_train.py 用 re-export 保留 sister script / test 对
`anima_train.X` 的访问路径不变。顺手删除死代码 `_lazy_imports()`（grep 0
caller）。

anima_train.py 行数 2901 → 2163；migration test 全绿。


* refactor(anima_train): 抽 models + sampling 到 training/（sister script 公开 API）

ADR 0003 PR-A 第 2 步：

- training/models.py    ← load_anima_model / load_vae / load_text_encoders /
                          ensure_models_namespace（被 anima_daemon / anima_generate /
                          anima_reg_ai 通过 _T.X 直接调用，PR-A re-export 保证 0 改动）
- training/sampling.py  ← sigma 调度（_time_snr_shift / _flow_sigmas_simple /
                          _default_noise_sampler / _sample_er_sde_const_x0）+
                          sample_image（训练/生成共用入口）

anima_train.py 行数 2163 → 1731；migration test 全绿。


* refactor(anima_train): 抽 dataset + collate + CLI 到 training/

ADR 0003 PR-A 第 3 步：

- training/dataset.py  ← BucketManager / ImageDataset / RepeatDataset /
                         MergedDataset / BucketBatchSampler /
                         CachedLatentDataset + collate_fn / collate_fn_cached
- training/cli.py      ← parse_args（被 test_anima_train_migration 直接调）+
                         _try_rich / _ask_* / _guess_default_paths / prompt_for_args

ImageDataset 里 caption_utils 动态加载的 `Path(__file__).parent`、
_guess_default_paths 里的 `Path(__file__).resolve().parent` 都加一层 .parent
保持等价语义（原指向 runtime/，搬到 runtime/training/ 后需多上一层）。注：
caption_utils 路径在原代码就指向不存在的 runtime/utils/，恒走 TXT 兜底；不在
本 PR 范畴内修。

anima_train.py 行数 1731 → 1042；migration test 全绿。


* refactor(anima_train): 抽 timestep_sampling / noise / loss_weighting 到 training/

ADR 0003 PR-A 第 4 步：3 个数学小模块，单文件多 fn 模式（不开 plugin
subfolder，每个 scheme 是同签名 ~30 行函数）：

- training/timestep_sampling.py — sample_t (logit_normal/uniform/mode/...)
- training/noise.py             — make_noise (offset + pyramid)
- training/loss_weighting.py    — compute_loss_weight (min_snr/cosmap/...)

anima_train.py 行数 1042 → 930；migration test 全绿。


* refactor(anima_train): thin shim 收尾——清死 import + 重写 docstring

ADR 0003 PR-A 第 5 步（收尾）：

- 改 module docstring：说明本文件已变 thin shim，实现层在 training/
- 删 main() 用不到的 top-level import：argparse / math / subprocess / types /
  Optional / Dataset
- 整理 re-export 段：加分隔注释，明确给 sister script + test 用，新代码请
  直接 from training.X import
- 清掉之前 commit 散落的"已搬到 ..."游离注释

anima_train.py 930 → 909；migration test 全绿，32 个 sister/test 用的公开
名字全部仍可 anima_train.X 访问。


---------
