+++ runtime/anima_train.py
@@ -1,34 +1,27 @@
 #!/usr/bin/env python
-"""
-Anima LoRA Trainer v2 - 支持 LyCORIS + 训练时推理
-基于 trainerV1.01 重构，轻量单文件
+"""Anima LoRA Trainer v2 — main() 编排入口。
 
-特性：
-- 标准 LoRA 和 LyCORIS LoKr 双模式
-- 训练时推理出图
-- Flow Matching 训练
-- ARB 分桶
-- 依赖自动检测与安装
-- Rich 进度条 + ASCII Loss 曲线
-- 梯度检查点支持
-- Caption 预处理 (shuffle/keep_tokens)
+本模块的实现层已按 ADR 0003 PR-A 拆到 runtime/training/ 子包：
+  bootstrap / cli / observability / model_loading / models / text_encoding /
+  state / dataset / sampling / timestep_sampling / noise / loss_weighting
+
+顶部的 re-export 段保留 anima_train.X 访问路径，给 sister script
+（anima_daemon / anima_generate / anima_reg_ai）和 tests/ 不变。新代码请
+直接 `from training.X import Y`。
+
+LoRA / LoKr 实现：见 utils.lycoris_adapter.AnimaLycorisAdapter（ADR 0001）。
 """
 
-import argparse
 import logging
-import math
 import os
 import random
-import subprocess
 import sys
 import time
-import types
 from pathlib import Path
-from typing import Optional
 
 # 脚本在 runtime/ 下按裸脚本启动（`python runtime/anima_train.py`）。
-# 把仓库根 + runtime/ 注入 sys.path，让 `import utils.*` / `import train_monitor` 等
-# 不需要改成包导入。
+# 把仓库根 + runtime/ 注入 sys.path，让 `import utils.*` / `import train_monitor` /
+# `import training.*` 等不需要改成包导入。
 _REPO_ROOT = Path(__file__).resolve().parent.parent
 for _p in (_REPO_ROOT, _REPO_ROOT / "runtime"):
     _ps = str(_p)
@@ -47,2055 +40,70 @@ for _stream in (sys.stdout, sys.stderr):
 
 import torch
 import torch.nn.functional as F
-from torch.utils.data import DataLoader, Dataset
+from torch.utils.data import DataLoader
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
 logger = logging.getLogger(__name__)
 
 
-# ============================================================================
-# 依赖检测
-# ============================================================================
-
-def ensure_dependencies(auto_install=False):
-    """检测并可选自动安装缺失依赖"""
-    required = {
-        "numpy": "numpy",
-        "PIL": "Pillow",
-        "safetensors": "safetensors",
-        "transformers": "transformers",
-        "einops": "einops",
-        "torchvision": "torchvision",
-        "yaml": "pyyaml",
-    }
-    missing = []
-    for module_name, pip_name in required.items():
-        try:
-            __import__(module_name)
-        except Exception:
-            missing.append(pip_name)
-    if not missing:
-        return
-    missing_list = ", ".join(sorted(set(missing)))
-    print(f"Missing dependencies: {missing_list}")
-    if not auto_install:
-        print(f"Install them with:\n  {sys.executable} -m pip install {missing_list}")
-        raise SystemExit(1)
-    cmd = [sys.executable, "-m", "pip", "install", *sorted(set(missing))]
-    print("Installing missing dependencies...")
-    try:
-        subprocess.run(cmd, check=False)
-    except Exception as exc:
-        print(f"Auto-install failed: {exc}")
-        raise SystemExit(1)
-    # Re-check after install
-    still_missing = []
-    for module_name, pip_name in required.items():
-        try:
-            __import__(module_name)
-        except Exception:
-            still_missing.append(pip_name)
-    if still_missing:
-        still_list = ", ".join(sorted(set(still_missing)))
-        print(f"Still missing: {still_list}")
-        raise SystemExit(1)
-
-
-# ============================================================================
-# YAML 配置加载
-# ============================================================================
-
-def load_yaml_config(config_path):
-    """加载 YAML 配置文件"""
-    try:
-        import yaml
-    except ImportError:
-        print("PyYAML not installed. Install with: pip install pyyaml")
-        raise SystemExit(1)
-
-    config_path = Path(config_path)
-    if not config_path.exists():
-        raise FileNotFoundError(f"Config file not found: {config_path}")
-
-    with open(config_path, "r", encoding="utf-8") as f:
-        config = yaml.safe_load(f)
-
-    if config is None:
-        config = {}
-
-    return config
-
-
-def apply_yaml_config(args, config):
-    """将 YAML 配置应用到 args；命令行显式参数优先于 YAML。
-
-    实现走 studio.argparse_bridge.merge_yaml_into_namespace —— 字段名 / 默认值
-    都从 studio.schema.TrainingConfig 这一份单一权威源派生，避免与 parse_args
-    脱节。未在 schema 中的 YAML 键会被忽略（拼写错误一目了然）。
-
-    在 merge 前调用 migrate_legacy_attention 兜底老 yaml 的 xformers/flash_attn
-    双 bool —— argparse_bridge 不走 pydantic validator，schema 层的迁移逻辑
-    无法生效，所以这里显式做一次。
-    """
-    from studio.argparse_bridge import merge_yaml_into_namespace
-    from studio.schema import TrainingConfig, migrate_legacy_attention
-    config = migrate_legacy_attention(dict(config or {}))
-    return merge_yaml_into_namespace(args, config, TrainingConfig)
-
-
-# Lazy imports after dependency check
-def _lazy_imports():
-    global np, Image
-    import numpy as np
-    from PIL import Image
-
-
-# ============================================================================
-# 进度和 Loss 曲线可视化
-# ============================================================================
-
-def init_progress(show_progress, total_steps):
-    """初始化 Rich 进度条"""
-    if not show_progress:
-        return None, None, None
-    try:
-        from rich.progress import (
-            BarColumn, MofNCompleteColumn, Progress, TextColumn,
-            TimeElapsedColumn, TimeRemainingColumn,
-        )
-        progress = Progress(
-            TextColumn("{task.description}"),
-            BarColumn(),
-            MofNCompleteColumn(),
-            TextColumn("loss={task.fields[loss]:.4f}"),
-            TextColumn("lr={task.fields[lr]:.2e}"),
-            TextColumn("speed={task.fields[speed]:.2f} it/s"),
-            TimeElapsedColumn(),
-            TimeRemainingColumn(),
-            refresh_per_second=10,
-        )
-        task = progress.add_task("train", total=total_steps, loss=0.0, lr=0.0, speed=0.0)
-        return progress, task, "rich"
-    except Exception:
-        return "plain", None, None
-
-
-def render_loss_curve(losses, width=60, height=10):
-    """渲染 ASCII Loss 曲线"""
-    if not losses:
-        return ""
-    if width < 5:
-        width = 5
-    values = losses
-    if len(values) > width:
-        step = len(values) / width
-        buckets = []
-        for i in range(width):
-            start = int(i * step)
-            end = int((i + 1) * step)
-            end = max(end, start + 1)
-            chunk = values[start:end]
-            buckets.append(sum(chunk) / len(chunk))
-        values = buckets
-    min_v = min(values)
-    max_v = max(values)
-    if max_v == min_v:
-        max_v = min_v + 1e-8
-    grid = [[" " for _ in range(len(values))] for _ in range(height)]
-    for i, v in enumerate(values):
-        y = int((v - min_v) / (max_v - min_v) * (height - 1))
-        y = height - 1 - y
-        grid[y][i] = "*"
-    lines = ["".join(row) for row in grid]
-    lines.append(f"min={min_v:.4f} max={max_v:.4f}")
-    return "\n".join(lines)
-
-
-def render_curve_panel(losses, width=60, height=10):
-    """渲染 Rich Panel 包装的 Loss 曲线"""
-    try:
-        from rich.panel import Panel
-        from rich.text import Text
-    except Exception:
-        return None
-    chart = render_loss_curve(losses, width=width, height=height)
-    return Panel(Text(chart), title="Loss curve (recent)", expand=False)
-
-
-# ============================================================================
-# Weights & Biases 可选监控
-# ============================================================================
-
-class WandBMonitor:
-    def __init__(
-        self,
-        wandb_module,
-        run,
-        *,
-        log_samples: bool = False,
-        sample_max_side: int = 512,
-        sample_every_n_steps: int = 0,
-    ) -> None:
-        self._wandb = wandb_module
-        self._run = run
-        self.log_samples = log_samples
-        self.sample_max_side = max(64, int(sample_max_side or 512))
-        self.sample_every_n_steps = max(0, int(sample_every_n_steps or 0))
-        self._last_logged_step: Optional[int] = None
-
-    @property
-    def enabled(self) -> bool:
-        return self._run is not None
-
-    def log(self, data: dict, *, step: Optional[int] = None) -> None:
-        if not self.enabled:
-            return
-        try:
-            self._run.log(data, step=step)
-        except Exception as exc:
-            logger.warning(f"W&B log 失败: {exc}")
-
-    def _should_log_step(self, key: str, step: Optional[int]) -> bool:
-        # baseline / epoch 边界一律放行；step 模式按 sample_every_n_steps 节流。
-        if self.sample_every_n_steps <= 0:
-            return True
-        if not key.startswith("samples/step"):
-            return True
-        if step is None or step <= 0:
-            return True
-        if step == self._last_logged_step:
-            return True  # 同步重复调用允许
-        return step % self.sample_every_n_steps == 0
-
-    def _prepare_image(self, image_path: Path, caption: str):
-        # 原图常 2K+，wandb 面板浏览 512px 已足够；JPEG 流量比 PNG 小一个数量级。
-        try:
-            from PIL import Image
-        except Exception:
-            return self._wandb.Image(str(image_path), caption=caption)
-        try:
-            with Image.open(image_path) as img:
-                img = img.convert("RGB")
-                max_side = self.sample_max_side
-                w, h = img.size
-                if max(w, h) > max_side:
-                    scale = max_side / float(max(w, h))
-                    new_size = (max(1, int(w * scale)), max(1, int(h * scale)))
-                    img = img.resize(new_size, Image.LANCZOS)
-                return self._wandb.Image(img, caption=caption)
-        except Exception as exc:
-            logger.warning(f"W&B 图片缩放失败，改传原图: {exc}")
-            return self._wandb.Image(str(image_path), caption=caption)
-
-    def log_image(self, key: str, image_path: Path, *, caption: str, step: Optional[int] = None) -> None:
-        if not self.enabled:
-            return
-        if not self._should_log_step(key, step):
-            return
-        try:
-            self._run.log({key: [self._prepare_image(image_path, caption)]}, step=step)
-            self._last_logged_step = step
-        except Exception as exc:
-            logger.warning(f"W&B 图片记录失败: {exc}")
-
-    def finish(self) -> None:
-        if not self.enabled:
-            return
-        try:
-            self._run.finish()
-        except Exception as exc:
-            logger.warning(f"W&B finish 失败: {exc}")
-
-
-def init_wandb_monitor(args, output_dir: Path, config_path: Optional[Path]) -> WandBMonitor:
-    enabled = str(os.environ.get("WANDB_ENABLED", "")).strip().lower()
-    if enabled not in {"1", "true", "yes", "on"}:
-        return WandBMonitor(None, None)
-    mode = str(os.environ.get("WANDB_MODE", "online") or "online")
-    if mode == "disabled":
-        return WandBMonitor(None, None)
-    try:
-        import wandb
-    except ImportError as exc:
-        raise RuntimeError(
-            "已在 Settings 启用 WandB，但当前环境没有安装 wandb。"
-            "请先在训练环境安装：pip install wandb，或在 Settings 关闭 WandB。"
-        ) from exc
-
-    project = os.environ.get("WANDB_PROJECT") or "AnimaLoraStudio"
-    entity = os.environ.get("WANDB_ENTITY") or None
-    run_name = os.environ.get("WANDB_RUN_NAME") or str(args.output_name)
-    # 默认关 — supervisor 已经按 secrets.wandb.log_samples 设过 env；env 缺省时也保持关闭。
-    log_samples = str(os.environ.get("WANDB_LOG_SAMPLES", "0")).strip().lower() not in {
-        "0", "false", "no", "off",
-    }
-    try:
-        sample_max_side = int(os.environ.get("WANDB_SAMPLE_MAX_SIDE", "512") or 512)
-    except ValueError:
-        sample_max_side = 512
-    try:
-        sample_every_n_steps = int(os.environ.get("WANDB_SAMPLE_EVERY_N_STEPS", "0") or 0)
-    except ValueError:
-        sample_every_n_steps = 0
-    wandb_dir = output_dir / "wandb"
-    wandb_dir.mkdir(parents=True, exist_ok=True)
-    cfg = {
-        key: value
-        for key, value in vars(args).items()
-        if key not in {"interactive", "auto_install"}
-    }
-    cfg["config_path"] = str(config_path) if config_path else ""
-    run = wandb.init(
-        project=project,
-        entity=entity,
-        name=run_name,
-        mode=mode,
-        config=cfg,
-        dir=str(wandb_dir),
-    )
-    logger.info(f"W&B 监控已启用: project={project}, run={run_name}, mode={mode}")
-    return WandBMonitor(
-        wandb,
-        run,
-        log_samples=log_samples,
-        sample_max_side=sample_max_side,
-        sample_every_n_steps=sample_every_n_steps,
-    )
-
-
-# ============================================================================
-# 梯度检查点
-# ============================================================================
-
-def forward_with_optional_checkpoint(model, latents, timesteps, cross, padding_mask, use_checkpoint=False):
-    """带可选梯度检查点的前向传播"""
-    if not use_checkpoint:
-        return model(latents, timesteps, cross, padding_mask=padding_mask)
-    from torch.utils.checkpoint import checkpoint
-
-    x_B_T_H_W_D, rope_emb, extra_pos_emb = model.prepare_embedded_sequence(
-        latents, fps=None, padding_mask=padding_mask,
-    )
-    if timesteps.ndim == 1:
-        timesteps = timesteps.unsqueeze(1)
-    t_embedding, adaln_lora = model.t_embedder(timesteps)
-    t_embedding = model.t_embedding_norm(t_embedding)
-
-    block_kwargs = {
-        "rope_emb_L_1_1_D": rope_emb,
-        "adaln_lora_B_T_3D": adaln_lora,
-        "extra_per_block_pos_emb": extra_pos_emb,
-    }
-
-    for block in model.blocks:
-        def custom_forward(x, blk=block):
-            return blk(x, t_embedding, cross, **block_kwargs)
-        x_B_T_H_W_D = checkpoint(custom_forward, x_B_T_H_W_D, use_reentrant=False)
-
-    x_B_T_H_W_O = model.final_layer(x_B_T_H_W_D, t_embedding, adaln_lora_B_T_3D=adaln_lora)
-    return model.unpatchify(x_B_T_H_W_O)
-
-
-# ============================================================================
-# xformers 支持
-# ============================================================================
-
-def enable_xformers(model):
-    """为模型启用 xformers memory efficient attention"""
-    try:
-        from xformers.ops import memory_efficient_attention
-    except ImportError:
-        logger.warning("xformers 未安装，跳过启用")
-        return False
-
-    enabled_count = 0
-    for name, module in model.named_modules():
-        # 查找 attention 模块并替换
-        if hasattr(module, "set_use_memory_efficient_attention_xformers"):
-            module.set_use_memory_efficient_attention_xformers(True)
-            enabled_count += 1
-        elif hasattr(module, "enable_xformers_memory_efficient_attention"):
-            module.enable_xformers_memory_efficient_attention()
-            enabled_count += 1
-
-    if enabled_count > 0:
-        logger.info(f"xformers 已启用: {enabled_count} 个模块")
-        return True
-
-    # 如果模型没有内置支持，尝试 monkey patch
-    logger.info("xformers 已加载，将在 attention 计算中使用")
-    return True
-
-
-# ============================================================================
-# 模型加载工具
-# ============================================================================
-
-def find_diffusion_pipe_root():
-    """查找 diffusion-pipe 模型代码路径。
-
-    候选顺序（首个命中即返回）：
-      1. 脚本同目录 `diffusion_models/` / `models/`（CLI 直接 cd 进 scripts/ 跑）
-      2. 仓库根 `models/` / `diffusion_models/`（训练脚本在 runtime/ 下的 repo
-         layout：repo_root/runtime/anima_train.py → ../models/anima_modeling.py）
-      3. 环境变量 `DIFFUSION_PIPE_ROOT`（覆盖路径用）
-    """
-    script_dir = Path(__file__).parent
-    repo_root = script_dir.parent
-    candidates = [
-        script_dir / "diffusion_models",
-        script_dir / "models",
-        repo_root / "models",
-        repo_root / "diffusion_models",
-        Path(os.environ.get("DIFFUSION_PIPE_ROOT", "")) if os.environ.get("DIFFUSION_PIPE_ROOT") else None,
-    ]
-    for candidate in candidates:
-        if candidate and (candidate / "anima_modeling.py").exists():
-            return candidate
-        if candidate and (candidate / "models" / "anima_modeling.py").exists():
-            return candidate / "models"
-    raise RuntimeError("找不到 anima_modeling.py，请设置 DIFFUSION_PIPE_ROOT 或放置模型代码")
-
-
-def load_module_from_path(module_name, file_path):
-    """动态加载 Python 模块"""
-    import importlib.util
-    spec = importlib.util.spec_from_file_location(module_name, file_path)
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
-
-
-def _strip_prefixes(key: str, prefixes: list[str]) -> str:
-    """反复剥离前缀（支持 module.model. 这种复合前缀）"""
-    if not prefixes:
-        return key
-    changed = True
-    while changed:
-        changed = False
-        for p in prefixes:
-            if key.startswith(p):
-                key = key[len(p) :]
-                changed = True
-    return key
-
-
-def _pick_best_prefix_remap(sd_keys: list[str], model_keys: set[str]) -> tuple[list[str], int]:
-    """
-    从常见前缀组合里选择“命中最多 model_keys”的 remap 方案。
-    返回 (prefixes, matched_count)。
-    """
-    candidates: list[tuple[str, list[str]]] = [
-        ("none", []),
-        ("net.", ["net."]),
-        ("model.", ["model."]),
-        ("module.", ["module."]),
-        ("module.+model.", ["module.", "model."]),
-        ("module.model.", ["module.model."]),
-        ("diffusion_model.", ["diffusion_model."]),
-        ("model.diffusion_model.", ["model.diffusion_model."]),
-        ("transformer.", ["transformer."]),
-        ("vae.", ["vae."]),
-        ("first_stage_model.", ["first_stage_model."]),
-        ("net.+model.", ["net.", "model."]),
-        ("net.model.", ["net.model."]),
-    ]
-
-    best_prefixes: list[str] = []
-    best_matched = -1
-    for _name, prefixes in candidates:
-        matched = 0
-        for k in sd_keys:
-            kk = _strip_prefixes(k, prefixes)
-            if kk in model_keys:
-                matched += 1
-        if matched > best_matched:
-            best_matched = matched
-            best_prefixes = prefixes
-    return best_prefixes, best_matched
-
-
-def _load_safetensors_state_dict(path: Path) -> dict:
-    from safetensors import safe_open
-
-    sd = {}
-    with safe_open(path, framework="pt", device="cpu") as f:
-        for k in f.keys():
-            sd[k] = f.get_tensor(k)
-    return sd
-
-
-def resolve_path_best_effort(path_str: str, bases: list[Path]) -> str:
-    """
-    将相对路径按多个 base 尝试解析到一个真实存在的路径。
-    主要用于：无论从 repo 根目录还是 AnimaLoraToolkit 目录启动，都能找到 models/* 文件。
-    """
-    if not path_str:
-        return path_str
-
-    p = Path(path_str)
-    if p.is_absolute():
-        return str(p)
-
-    # 先按原样（相对 cwd）试一下
-    if p.exists():
-        return str(p)
-
-    # 逐 base 拼接尝试
-    for b in bases:
-        if not b:
-            continue
-        try:
-            cand = (Path(b) / p).resolve()
-        except Exception:
-            cand = Path(b) / p
-        if cand.exists():
-            return str(cand)
-
-    # 常见：配置写了 AnimaLoraToolkit/xxx，但启动目录已经在 AnimaLoraToolkit 下
-    parts = p.parts
-    if parts and parts[0].lower() in ("animaloratoolkit", "anima_trainer", "anima-trainer"):
-        p2 = Path(*parts[1:])
-        if p2.exists():
-            return str(p2)
-        for b in bases:
-            if not b:
-                continue
-            cand = Path(b) / p2
-            if cand.exists():
-                return str(cand)
-
-    return path_str
-
-
-def _load_weights_best_effort(model: torch.nn.Module, sd: dict, label: str) -> dict:
-    """
-    更健壮的权重加载：
-    - 自动尝试剥离常见前缀（model./module./...）
-    - 打印匹配率、missing/unexpected
-    - 关键模块未加载时直接报错（避免“采样全噪点”还继续训练）
-    """
-    model_keys = set(model.state_dict().keys())
-    sd_keys = list(sd.keys())
-    prefixes, matched = _pick_best_prefix_remap(sd_keys, model_keys)
-    remapped = {_strip_prefixes(k, prefixes): v for k, v in sd.items()}
-
-    incompatible = model.load_state_dict(remapped, strict=False)
-    missing = list(getattr(incompatible, "missing_keys", []) or [])
-    unexpected = list(getattr(incompatible, "unexpected_keys", []) or [])
-
-    matched_after = len(set(remapped.keys()) & model_keys)
-    coverage = matched_after / max(1, len(model_keys))
-    remap_name = "+".join(prefixes) if prefixes else "none"
-
-    logger.info(
-        f"{label} 权重加载: remap={remap_name}, 匹配 {matched_after}/{len(model_keys)} ({coverage:.1%}), "
-        f"missing={len(missing)}, unexpected={len(unexpected)}"
-    )
-
-    # 关键层缺失会直接导致输出接近 0，采样就是纯噪点
-    critical_prefixes = ("x_embedder.", "blocks.", "final_layer.")
-    critical_missing = [k for k in missing if k.startswith(critical_prefixes)]
-    if coverage < 0.60 or len(critical_missing) > 0:
-        preview_missing = ", ".join(critical_missing[:8])
-        raise RuntimeError(
-            f"{label} 权重看起来没有正确加载（remap={remap_name}, coverage={coverage:.1%}）。"
-            f"关键参数缺失: {preview_missing or 'N/A'}。\n"
-            f"这通常表示你选错了 .safetensors（不是完整 transformer/vae 权重），或 checkpoint key 前缀不匹配。"
-        )
-    return {
-        "remap": remap_name,
-        "coverage": coverage,
-        "missing": missing,
-        "unexpected": unexpected,
-    }
-
-
-def ensure_models_namespace(repo_root):
-    """确保 models 命名空间可用"""
-    repo_root = Path(repo_root)
-    if str(repo_root) not in sys.path:
-        sys.path.insert(0, str(repo_root))
-    if str(repo_root.parent) not in sys.path:
-        sys.path.insert(0, str(repo_root.parent))
-
-
-def load_anima_model(transformer_path, device, dtype, repo_root, *, flash_attn: bool = True):
-    """加载 Anima transformer 模型。
-
-    `flash_attn=False` 显式禁用 flash_attn fast path（attention_backend=xformers/none
-    时由 caller 传入），让 caller 完全决定 attention 实现 —— PR #17 那版默认
-    fn(True) 强制开 flash_attn 不让用户关，与 cfg.attention_backend 解耦不彻底。
-    """
-    from safetensors import safe_open
-
-    ensure_models_namespace(repo_root)
-
-    # 加载模型类
-    cosmos_modeling = load_module_from_path(
-        "cosmos_predict2_modeling",
-        repo_root / "cosmos_predict2_modeling.py",
-    )
-    anima_modeling = load_module_from_path(
-        "anima_modeling",
-        repo_root / "anima_modeling.py",
-    )
-    Anima = anima_modeling.Anima
-
-    # flash_attn 全局开关：set_flash_attn_enabled 内部检查 _FLASH_ATTN_AVAILABLE，
-    # 未装时返回 False 不抛错（idempotent）。用 caller 传入的 flash_attn 而不是
-    # 强制 True，让 attention_backend=none/xformers 时显式关掉。
-    fn = getattr(cosmos_modeling, "set_flash_attn_enabled", None)
-    if fn is not None:
-        try:
-            if fn(flash_attn):
-                logger.info("flash_attn 启用（训练 + sample 走 fast path）")
-            else:
-                logger.info("flash_attn 关闭（attention_backend=%s 或包未安装）",
-                            "flash_attn" if flash_attn else "non-flash")
-        except Exception as exc:  # noqa: BLE001
-            logger.warning("flash_attn 启用失败，继续走 SDPA fallback: %s", exc)
-
-    # 从 checkpoint 推断配置
-    with safe_open(transformer_path, framework="pt", device="cpu") as f:
-        for k in f.keys():
-            if k.endswith("x_embedder.proj.1.weight"):
-                w = f.get_tensor(k)
-                break
-
-    in_channels = (w.shape[1] // 4) - 1  # concat_padding_mask=True
-    model_channels = w.shape[0]
-
-    if model_channels == 2048:
-        num_blocks, num_heads = 28, 16
-    elif model_channels == 5120:
-        num_blocks, num_heads = 36, 40
-    else:
-        raise RuntimeError(f"未知的 model_channels={model_channels}")
-
-    config = dict(
-        max_img_h=240, max_img_w=240, max_frames=128,
-        in_channels=in_channels, out_channels=16,
-        patch_spatial=2, patch_temporal=1,
-        concat_padding_mask=True,
-        model_channels=model_channels,
-        num_blocks=num_blocks, num_heads=num_heads,
-        crossattn_emb_channels=1024,
-        pos_emb_cls="rope3d", pos_emb_learnable=True,
-        pos_emb_interpolation="crop",
-        use_adaln_lora=True, adaln_lora_dim=256,
-        rope_h_extrapolation_ratio=4.0 if in_channels == 16 else 3.0,
-        rope_w_extrapolation_ratio=4.0 if in_channels == 16 else 3.0,
-        rope_t_extrapolation_ratio=1.0,
-    )
-
-    model = Anima(**config)
-
-    # 加载权重
-    sd = _load_safetensors_state_dict(Path(transformer_path))
-    info = _load_weights_best_effort(model, sd, label="Transformer")
-
-    # 如果 checkpoint 中完全没有 llm_adapter 权重，随机初始化会把 cross-attn 条件搞乱，直接禁用更安全
-    has_llm_adapter = any("llm_adapter" in k for k in sd.keys())
-    if not has_llm_adapter and hasattr(model, "llm_adapter"):
-        try:
-            model.llm_adapter = None
-            logger.warning("检测到 checkpoint 不包含 llm_adapter 权重：已禁用 llm_adapter（回退为直接使用 Qwen embeddings）")
-        except Exception:
-            pass
-    model = model.to(device=device, dtype=dtype)
-    model.requires_grad_(False)
-
-    logger.info(f"Anima 模型加载完成: {model_channels}ch, {num_blocks} blocks")
-    return model
-
-
-def load_vae(vae_path, device, dtype, repo_root):
-    """加载 VAE"""
-    wan_vae = load_module_from_path("wan_vae", repo_root / "wan" / "vae2_1.py")
-    WanVAE = wan_vae.WanVAE_
-
-    cfg = dict(
-        dim=96, z_dim=16, dim_mult=[1, 2, 4, 4],
-        num_res_blocks=2, attn_scales=[],
-        temperal_downsample=[False, True, True], dropout=0.0,
-    )
-
-    model = WanVAE(**cfg).eval().requires_grad_(False)
-
-    sd = _load_safetensors_state_dict(Path(vae_path))
-    _load_weights_best_effort(model, sd, label="VAE")
-    model = model.to(device=device, dtype=dtype)
-
-    # VAE 归一化参数
-    mean = torch.tensor([
-        -0.7571, -0.7089, -0.9113, 0.1075, -0.1745, 0.9653, -0.1517, 1.5508,
-        0.4134, -0.0715, 0.5517, -0.3632, -0.1922, -0.9497, 0.2503, -0.2921
-    ], dtype=dtype, device=device)
-    std = torch.tensor([
-        2.8184, 1.4541, 2.3275, 2.6558, 1.2196, 1.7708, 2.6052, 2.0743,
-        3.2687, 2.1526, 2.8652, 1.5579, 1.6382, 1.1253, 2.8251, 1.9160
-    ], dtype=dtype, device=device)
-
-    class VAEWrapper:
-        pass
-
-    wrapper = VAEWrapper()
-    wrapper.model = model
-    wrapper.mean = mean
-    wrapper.std = std
-    wrapper.scale = [mean, 1.0 / std]
-
-    logger.info("VAE 加载完成")
-    return wrapper
-
-
-def load_text_encoders(qwen_path, t5_tokenizer_path, device, dtype):
-    """加载文本编码器"""
-    from transformers import AutoModelForCausalLM, AutoTokenizer, T5Tokenizer
-
-    # Qwen
-    qwen_tokenizer = AutoTokenizer.from_pretrained(qwen_path, trust_remote_code=True)
-    qwen_model = AutoModelForCausalLM.from_pretrained(
-        qwen_path, torch_dtype=dtype, trust_remote_code=True
-    ).to(device).eval().requires_grad_(False)
-
-    # T5 tokenizer
-    if t5_tokenizer_path and Path(t5_tokenizer_path).exists():
-        t5_tokenizer = T5Tokenizer.from_pretrained(t5_tokenizer_path)
-    else:
-        t5_tokenizer = T5Tokenizer.from_pretrained("google/t5-v1_1-xxl")
-
-    logger.info("文本编码器加载完成")
-    return qwen_model, qwen_tokenizer, t5_tokenizer
-
-
-# ============================================================================
-# 文本编码
-# ============================================================================
-
-def encode_qwen(model, tokenizer, texts, device, max_length=512):
-    """Qwen 文本编码"""
-    # Qwen3 tokenizer 对空字符串可能返回 0 tokens（会导致模型内部 reshape 失败）
-    # ComfyUI 的 AnimaTokenizer 设置了 min_length=1，这里做同等兜底。
-    if isinstance(texts, str):
-        texts = [texts]
-    texts = [(" " if (t is None or str(t).strip() == "") else str(t)) for t in texts]
-
-    inputs = tokenizer(
-        texts,
-        return_tensors="pt",
-        padding=True,
-        truncation=True,
-        max_length=max_length,
-        add_special_tokens=False,
-    )
-    # 仍可能出现空序列（极端 tokenizer 行为），强制塞 1 个 token
-    if inputs["input_ids"].ndim == 2 and inputs["input_ids"].shape[1] == 0:
-        pad_id = tokenizer.pad_token_id
-        if pad_id is None:
-            pad_id = tokenizer.eos_token_id
-        if pad_id is None:
-            pad_id = 0
-        bs = len(texts)
-        inputs["input_ids"] = torch.full((bs, 1), int(pad_id), dtype=torch.long)
-        inputs["attention_mask"] = torch.ones((bs, 1), dtype=torch.long)
-    inputs = inputs.to(device)
-
-    with torch.inference_mode():
-        outputs = model(
-            input_ids=inputs["input_ids"],
-            attention_mask=inputs["attention_mask"],
-            output_hidden_states=True,
-            return_dict=True,
-            use_cache=False,
-        )
-
-    hidden = outputs.hidden_states[-1]
-    # 清零 padding 位置
-    mask = inputs["attention_mask"].unsqueeze(-1)
-    hidden = hidden * mask
-
-    return hidden, inputs["attention_mask"]
-
-
-def _time_snr_shift(alpha: float, t: torch.Tensor) -> torch.Tensor:
-    """ComfyUI ModelSamplingDiscreteFlow.time_snr_shift"""
-    if alpha == 1.0:
-        return t
-    return alpha * t / (1 + (alpha - 1) * t)
-
-
-def _flow_sigmas_simple(steps: int, *, shift: float = 3.0, timesteps: int = 1000, device: str = "cpu") -> torch.Tensor:
-    """
-    复刻 ComfyUI:
-    - supported_models.Anima 的 sampling_settings: shift=3.0, multiplier=1.0
-    - ModelSamplingDiscreteFlow + simple_scheduler(model_sampling, steps)
-
-    返回：sigmas (steps+1,) float32，从高到低，末尾带 0.0
-    """
-    ts = torch.arange(1, timesteps + 1, device=device, dtype=torch.float32) / float(timesteps)  # (0, 1]
-    sigmas_full = _time_snr_shift(float(shift), ts)  # (0, 1]
-
-    ss = len(sigmas_full) / float(steps)
-    sigmas = [float(sigmas_full[-(1 + int(i * ss))]) for i in range(steps)]
-    sigmas.append(0.0)
-    sigmas = torch.tensor(sigmas, device=device, dtype=torch.float32)
-
-    # ComfyUI offset_first_sigma_for_snr: CONST 下避免 sigma=1 导致 logit inf
-    if sigmas.numel() > 0 and sigmas[0] >= 1.0:
-        sigmas[0] = float(_time_snr_shift(float(shift), torch.tensor(1.0 - 1e-4, device=device, dtype=torch.float32)))
-    return sigmas
-
-
-def _default_noise_sampler(x: torch.Tensor, seed: int | None):
-    """参考 ComfyUI k_diffusion_sampling.default_noise_sampler"""
-    if seed is not None:
-        if x.device.type == "cpu":
-            seed = int(seed) + 1
-        g = torch.Generator(device=x.device)
-        g.manual_seed(int(seed))
-    else:
-        g = None
-
-    def _sample(_sigma, _sigma_next):
-        return torch.randn(x.size(), dtype=x.dtype, layout=x.layout, device=x.device, generator=g)
-
-    return _sample
-
-
-@torch.no_grad()
-def _sample_er_sde_const_x0(
-    denoise_fn,
-    x: torch.Tensor,
-    sigmas: torch.Tensor,
-    *,
-    seed: int | None = None,
-    s_noise: float = 1.0,
-    max_stage: int = 3,
-    step_callback=None,
-):
-    """
-    Extended Reverse-Time SDE solver（ER-SDE-Solver-3）在 CONST(flow) 噪声日程下的实现。
-    参考 ComfyUI 的 k_diffusion_sampling.sample_er_sde（删去 model_patcher 依赖）。
-
-    step_callback：可选钩子（仅 daemon 中间步预览用，commit 14）。签名
-        callback(step:int, total:int, denoised:torch.Tensor) → None。每步算
-        完 x0 估计（denoised）调一次；同步阻塞返回 —— 调用方应做轻量解码 +
-        异步 push，不在 callback 内阻塞。默认 None 时行为完全等价旧版。
-    """
-    sigmas = sigmas.to(device=x.device, dtype=torch.float32)
-    if sigmas.numel() <= 1:
-        return x
-
-    noise_sampler = _default_noise_sampler(x, seed=seed)
-
-    # CONST: half_log_snr = log((1 - t) / t) = -logit(t)
-    eps = 1e-12
-    t = sigmas.clamp(min=eps, max=1.0 - eps)
-    half_log_snrs = torch.log((1 - t) / t)
-    er_lambdas = half_log_snrs.neg().exp()  # er_lambda = t / (1 - t)
-
-    old_denoised = None
-    old_denoised_d = None
-
-    def noise_scaler(lam: torch.Tensor) -> torch.Tensor:
-        # default_er_sde_noise_scaler
-        lam = lam.to(x.device, dtype=torch.float32)
-        return lam * ((lam ** 0.3).exp() + 10.0)
-
-    num_integration_points = 200.0
-    point_indice = torch.arange(0, num_integration_points, dtype=torch.float32, device=x.device)
-
-    for i in range(len(sigmas) - 1):
-        sigma = sigmas[i]
-        denoised = denoise_fn(x, sigma)
-
-        if step_callback is not None:
-            try:
-                step_callback(i, len(sigmas) - 1, denoised)
-            except Exception:
-                pass  # 预览失败不该影响采样
-
-        stage_used = min(int(max_stage), i + 1)
-        if sigmas[i + 1] == 0:
-            x = denoised
-        else:
-            er_lambda_s, er_lambda_t = er_lambdas[i], er_lambdas[i + 1]
-            alpha_s = 1.0 - sigmas[i]
-            alpha_t = 1.0 - sigmas[i + 1]
-            r_alpha = alpha_t / alpha_s
-            r = noise_scaler(er_lambda_t) / noise_scaler(er_lambda_s)
-
-            # Stage 1 (Euler)
-            x = r_alpha * r * x + alpha_t * (1 - r) * denoised
-
-            if stage_used >= 2 and old_denoised is not None:
-                dt = er_lambda_t - er_lambda_s
-                lambda_step_size = -dt / num_integration_points
-                lambda_pos = er_lambda_t + point_indice * lambda_step_size
-                scaled_pos = noise_scaler(lambda_pos)
-
-                # Stage 2
-                s = torch.sum(1 / scaled_pos) * lambda_step_size
-                denoised_d = (denoised - old_denoised) / (er_lambda_s - er_lambdas[i - 1])
-                x = x + alpha_t * (dt + s * noise_scaler(er_lambda_t)) * denoised_d
-
-                if stage_used >= 3 and old_denoised_d is not None:
-                    # Stage 3
-                    s_u = torch.sum((lambda_pos - er_lambda_s) / scaled_pos) * lambda_step_size
-                    denoised_u = (denoised_d - old_denoised_d) / ((er_lambda_s - er_lambdas[i - 2]) / 2)
-                    x = x + alpha_t * ((dt ** 2) / 2 + s_u * noise_scaler(er_lambda_t)) * denoised_u
-
-                old_denoised_d = denoised_d
-
-            # Stochastic term
-            if s_noise and float(s_noise) > 0:
-                noise = noise_sampler(float(sigmas[i]), float(sigmas[i + 1]))
-                sde_scale = (er_lambda_t ** 2 - (er_lambda_s ** 2) * (r ** 2)).clamp(min=0).sqrt().nan_to_num(nan=0.0)
-                x = x + alpha_t * noise * float(s_noise) * sde_scale
-
-        old_denoised = denoised
-
-    return x
-
-
-def _parse_weighted_tag(tag: str) -> tuple[str, float]:
-    """
-    解析单个 tag 的权重（参考指南“权重控制”）。
-    支持：
-    - (tag:1.5)
-    - (tag) / ((tag))  => 1.1^n
-    - [tag]            => 1/1.1
-    """
-    import re
-
-    s = tag.strip()
-    if not s:
-        return "", 1.0
-
-    # 显式 (xxx:1.23)
-    m = re.fullmatch(r"\(\s*(.+?)\s*:\s*([+-]?\d+(?:\.\d+)?)\s*\)", s)
-    if m:
-        return m.group(1).strip(), float(m.group(2))
-
-    # 统计外层 () / [] 深度
-    w = 1.0
-    while True:
-        s2 = s.strip()
-        if len(s2) >= 2 and s2[0] == "(" and s2[-1] == ")":
-            s = s2[1:-1].strip()
-            w *= 1.1
-            continue
-        if len(s2) >= 2 and s2[0] == "[" and s2[-1] == "]":
-            s = s2[1:-1].strip()
-            w /= 1.1
-            continue
-        break
-    return s.strip(), float(w)
-
-
-def _build_qwen_text_from_prompt(prompt: str) -> str:
-    # Qwen 通道不传权重，只传“干净标签文本”（参考 ComfyUI anima-kai 的做法）
-    parts = [p.strip() for p in prompt.split(",") if p.strip()]
-    clean = []
-    for p in parts:
-        t, _w = _parse_weighted_tag(p)
-        if t:
-            clean.append(t)
-    return ", ".join(clean)
-
-
-def tokenize_t5_weighted(tokenizer, texts, max_length=512):
-    """
-    参考 ComfyUI 的 anima-kai：按逗号切分 tag，逐 tag 分词，并为每个 token 附带权重。
-    返回：input_ids, attention_mask(1=有效), token_weights
-    """
-    import torch
-
-    if isinstance(texts, str):
-        texts = [texts]
-
-    pad_id = tokenizer.pad_token_id if tokenizer.pad_token_id is not None else 0
-    eos_id = tokenizer.eos_token_id if tokenizer.eos_token_id is not None else 1
-
-    all_ids = []
-    all_w = []
-    for text in texts:
-        tags = [t.strip() for t in str(text).split(",") if t.strip()]
-        ids = []
-        ws = []
-        for tag in tags:
-            clean_tag, weight = _parse_weighted_tag(tag)
-            if not clean_tag:
-                continue
-            tok = tokenizer(clean_tag, add_special_tokens=False)
-            for tid in tok["input_ids"]:
-                ids.append(int(tid))
-                ws.append(float(weight))
-
-        # 末尾补一个 eos（ComfyUI 也是最后加一个终止 token）
-        ids.append(int(eos_id))
-        ws.append(1.0)
-
-        # 截断到 max_length（保留最后一个 eos）
-        if max_length and len(ids) > max_length:
-            ids = ids[: max_length - 1] + [int(eos_id)]
-            ws = ws[: max_length - 1] + [1.0]
-
-        all_ids.append(torch.tensor(ids, dtype=torch.long))
-        all_w.append(torch.tensor(ws, dtype=torch.float32))
-
-    # pad 到 batch 内最长
-    max_len = max(x.numel() for x in all_ids) if all_ids else 1
-    input_ids = torch.full((len(all_ids), max_len), pad_id, dtype=torch.long)
-    token_w = torch.zeros((len(all_w), max_len), dtype=torch.float32)
-    attention_mask = torch.zeros((len(all_ids), max_len), dtype=torch.long)
-
-    for i, (ids, ws) in enumerate(zip(all_ids, all_w)):
-        L = ids.numel()
-        input_ids[i, :L] = ids
-        token_w[i, :L] = ws
-        attention_mask[i, :L] = 1
-
-    return input_ids, attention_mask, token_w
-
-
-# ============================================================================
-# LoRA / LoKr 实现：见 utils.lycoris_adapter.AnimaLycorisAdapter
-# （历史自实现版本于 Stage 3c 删除，使用 lycoris-lora 包替代）
-# ============================================================================
-
-# ============================================================================
-# 训练状态保存/恢复（断点续训）
-# ============================================================================
-
-def save_training_state(path, injector, optimizer, epoch, global_step, loss_history=None, rng_state=None, monitor_state=None, scheduler=None):
-    """保存完整训练状态，支持断点续训"""
-    state = {
-        "lora_state_dict": injector.state_dict(),
-        "optimizer_state_dict": optimizer.state_dict(),
-        "epoch": epoch,
-        "global_step": global_step,
-        "loss_history": loss_history or [],
-        "rng_state": {
-            "torch": torch.get_rng_state(),
-            "cuda": torch.cuda.get_rng_state() if torch.cuda.is_available() else None,
-            "random": random.getstate(),
-        },
-        "monitor_state": monitor_state,  # 保存监控面板数据（用于恢复 loss 曲线）
-    }
-    if scheduler is not None:
-        state["scheduler_state_dict"] = scheduler.state_dict()
-    torch.save(state, path)
-    logger.info(f"训练状态已保存: {path} (epoch={epoch}, step={global_step})")
-
-
-def load_training_state(path, injector, optimizer, scheduler=None):
-    """加载训练状态，返回 (epoch, global_step, loss_history, monitor_state)"""
-    logger.info(f"加载训练状态: {path}")
-    state = torch.load(path, map_location="cpu", weights_only=False)
-    
-    # 加载 LoRA 权重（lycoris-lora backend）— 一次性导入 state_dict
-    # 旧自实现 ckpt 在 Stage 4 plan 决策中**不做迁移**，strict=False 让缺失键
-    # 走默认初始化路径而非崩溃；用户应当从头训练新格式 ckpt。
-    lora_sd = state["lora_state_dict"]
-    result = injector.load_state_dict(lora_sd, strict=False)
-    missing = len(getattr(result, "missing_keys", [])) if hasattr(result, "missing_keys") else 0
-    unexpected = len(getattr(result, "unexpected_keys", [])) if hasattr(result, "unexpected_keys") else 0
-    if missing or unexpected:
-        logger.warning(
-            f"resume LoRA: missing={missing}, unexpected={unexpected}（旧格式 ckpt？）"
-        )
-    
-    # 加载优化器状态
-    optimizer.load_state_dict(state["optimizer_state_dict"])
-
-    # 加载调度器状态
-    if scheduler is not None and "scheduler_state_dict" in state:
-        try:
-            scheduler.load_state_dict(state["scheduler_state_dict"])
-        except Exception as e:
-            logger.warning(f"调度器状态恢复失败（将从头开始）: {e}")
-
-    # 恢复随机数状态
-    if "rng_state" in state:
-        rng = state["rng_state"]
-        if rng.get("torch") is not None:
-            torch.set_rng_state(rng["torch"])
-        if rng.get("cuda") is not None and torch.cuda.is_available():
-            torch.cuda.set_rng_state(rng["cuda"])
-        if rng.get("random") is not None:
-            random.setstate(rng["random"])
-    
-    epoch = state.get("epoch", 0)
-    global_step = state.get("global_step", 0)
-    loss_history = state.get("loss_history", [])
-    monitor_state = state.get("monitor_state", None)  # 恢复监控数据
-    
-    logger.info(f"训练状态已恢复: epoch={epoch}, step={global_step}")
-    return epoch, global_step, loss_history, monitor_state
-
-
-# ============================================================================
-# 数据集
-# ============================================================================
-
-class BucketManager:
-    """ARB 分桶管理"""
-    def __init__(self, base_reso=1024, min_reso=512, max_reso=2048, step=64):
-        self.base_reso = base_reso
-        self.buckets = self._generate(min_reso, max_reso, step, base_reso)
-
-    def _generate(self, min_r, max_r, step, base):
-        buckets = []
-        base_area = base * base
-        for w in range(min_r, max_r + 1, step):
-            for h in range(min_r, max_r + 1, step):
-                if abs(w * h - base_area) / base_area > 0.1:
-                    continue
-                if max(w/h, h/w) > 2.0:
-                    continue
-                buckets.append((w, h))
-        return buckets
-
-    def get_bucket(self, w, h):
-        aspect = w / h
-        best = (self.base_reso, self.base_reso)
-        best_diff = float("inf")
-        for bw, bh in self.buckets:
-            diff = abs(aspect - bw/bh)
-            if diff < best_diff:
-                best_diff = diff
-                best = (bw, bh)
-        return best
-
-
-class ImageDataset(Dataset):
-    """
-    图像数据集
-    
-    支持两种 caption 格式：
-    1. JSON 文件（优先）- 支持分类 shuffle
-    2. TXT 文件（回退）- 传统 shuffle
-    """
-    # 保持与 studio/datasets.py:IMAGE_EXTS 同步（anima_train.py 是独立 CLI 脚本，
-    # 不强制 import studio package；改一处时另一处也要跟着改）。
-    EXTS = {".jpg", ".jpeg", ".png", ".webp", ".bmp", ".gif"}
-
-    def __init__(self, data_dir, resolution=1024, bucket_mgr=None,
-                 shuffle_caption=False, keep_tokens=0, flip_augment=False,
-                 tag_dropout=0.0, prefer_json=True, caption_override=None):
-        self.data_dir = Path(data_dir)
-        self.resolution = resolution
-        self.bucket_mgr = bucket_mgr
-        self.shuffle_caption = shuffle_caption
-        self.keep_tokens = keep_tokens
-        self.flip_augment = flip_augment
-        self.tag_dropout = tag_dropout
-        self.prefer_json = prefer_json
-        self.caption_override = caption_override  # 正则集：统一 caption，如 "1girl, solo"
-        
-        # 尝试导入 caption_utils（直接导入避开 __init__.py）
-        self.caption_utils = None
-        if prefer_json:
-            try:
-                import importlib.util
-                import sys
-                
-                # 直接加载 caption_utils.py
-                utils_path = Path(__file__).parent / "utils" / "caption_utils.py"
-                if utils_path.exists():
-                    spec = importlib.util.spec_from_file_location("caption_utils", utils_path)
-                    caption_module = importlib.util.module_from_spec(spec)
-                    sys.modules["caption_utils"] = caption_module
-                    spec.loader.exec_module(caption_module)
-                    
-                    self.caption_utils = {
-                        "load_and_build": caption_module.load_and_build_caption,
-                        "load_json": caption_module.load_caption_json,
-                        "normalize": caption_module.normalize_caption_json,
-                        "build": caption_module.build_caption_from_json,
-                    }
-                    logger.info("JSON caption 模式已启用（分类 shuffle）")
-                else:
-                    logger.warning(f"caption_utils.py 未找到: {utils_path}")
-            except Exception as e:
-                logger.warning(f"caption_utils 加载失败: {e}，回退到 TXT 模式")
-        
-        self.samples = self._scan()
-        json_count = sum(1 for s in self.samples if s.get("json_path"))
-        txt_count = len(self.samples) - json_count
-        unique_count = len(set(id(s) for s in self.samples))
-        logger.info(f"数据集: {unique_count} 张图 → {len(self.samples)} 样本（含 repeat）(JSON: {json_count}, TXT: {txt_count})")
-
-    @staticmethod
-    def _parse_repeats_from_dir(name: str) -> int:
-        """从文件夹名解析 Kohya 风格重复次数，如 '5_concept' → 5"""
-        prefix = name.split("_", 1)[0]
-        if prefix.isdigit():
-            return max(int(prefix), 1)
-        return 1
-
-    def _make_sample(self, img_path):
-        """为单张图构建 sample dict，找不到 caption 返回 None"""
-        sample = {"image": img_path}
-        json_path = img_path.with_suffix(".json")
-        if self.prefer_json and json_path.exists():
-            sample["json_path"] = json_path
-            sample["txt_path"] = None
-        else:
-            txt_path = img_path.with_suffix(".txt")
-            if not txt_path.exists():
-                txt_path = img_path.with_suffix(".caption")
-            if not txt_path.exists():
-                return None
-            sample["json_path"] = None
-            sample["txt_path"] = txt_path
-        return sample
-
-    def _scan(self):
-        """扫描数据集目录，支持 Kohya 风格文件夹重复。
-
-        目录结构示例::
-
-            dataset/
-            ├── 1_old/       ← repeat 1
-            │   ├── img.jpg
-            │   └── img.txt
-            └── 5_new/       ← repeat 5
-                ├── img.jpg
-                └── img.txt
-
-        没有数字前缀的文件夹或根目录下的图片按 repeat=1 处理。
-        """
-        unique_samples = []
-        folder_info = []  # (folder_name, repeat, count) for logging
-
-        # 收集根目录下的图片（repeat=1）
-        root_count = 0
-        for p in sorted(self.data_dir.iterdir()):
-            if p.is_file() and p.suffix.lower() in self.EXTS:
-                s = self._make_sample(p)
-                if s:
-                    s["_repeat"] = 1
-                    unique_samples.append(s)
-                    root_count += 1
-        if root_count:
-            folder_info.append(("(root)", 1, root_count))
-
-        # 收集子文件夹中的图片（解析 repeat）
-        for subdir in sorted(self.data_dir.iterdir()):
-            if not subdir.is_dir():
-                continue
-            repeats = self._parse_repeats_from_dir(subdir.name)
-            count = 0
-            for img_path in sorted(subdir.rglob("*")):
-                if img_path.suffix.lower() not in self.EXTS:
-                    continue
-                s = self._make_sample(img_path)
-                if s:
-                    s["_repeat"] = repeats
-                    unique_samples.append(s)
-                    count += 1
-            if count:
-                folder_info.append((subdir.name, repeats, count))
-
-        # 展开 repeat：将每个样本按其 repeat 次数复制
-        samples = []
-        for s in unique_samples:
-            r = s.pop("_repeat", 1)
-            for _ in range(r):
-                samples.append(s)
-
-        # 日志：每个文件夹的 repeat 信息
-        if folder_info:
-            for name, rep, cnt in folder_info:
-                logger.info(f"  文件夹 {name}: {cnt} 张 × repeat {rep} = {cnt * rep} 样本")
-
-        return samples
-
-    def _process_caption_txt(self, caption):
-        """处理 TXT caption: 传统 tag 打乱 + keep_tokens"""
-        if not caption:
-            return ""
-        if "," in caption:
-            tags = [t.strip() for t in caption.split(",")]
-        else:
-            tags = caption.split()
-
-        if self.keep_tokens > 0:
-            kept = tags[:self.keep_tokens]
-            rest = tags[self.keep_tokens:]
-            if self.shuffle_caption:
-                random.shuffle(rest)
-            tags = kept + rest
-        elif self.shuffle_caption:
-            random.shuffle(tags)
-
-        return ", ".join(tags)
-
-    def _process_caption_json(self, json_path):
-        """处理 JSON caption: 分类 shuffle"""
-        if self.caption_utils is None:
-            return None
-        
-        try:
-            raw_json = self.caption_utils["load_json"](json_path)
-            if raw_json is None:
-                return None
-            
-            # 检查是否已经是标准格式
-            if "tags" in raw_json and "meta" in raw_json:
-                normalized = raw_json
-            else:
-                normalized = self.caption_utils["normalize"](raw_json)
-            
-            # 构建 caption（分类 shuffle）
-            return self.caption_utils["build"](
-                normalized,
-                shuffle_appearance=self.shuffle_caption,
-                shuffle_tags=self.shuffle_caption,
-                shuffle_environment=self.shuffle_caption,
-                tag_dropout=self.tag_dropout,
-            )
-        except Exception as e:
-            logger.warning(f"JSON 处理失败 {json_path}: {e}")
-            return None
-
-    def __len__(self):
-        return len(self.samples)
-
-    def __getitem__(self, idx):
-        import numpy as np
-        from PIL import Image
-        sample = self.samples[idx]
-        img = Image.open(sample["image"]).convert("RGB")
-        
-        # 获取 caption（正则集可用 caption_override 统一覆盖）
-        caption = None
-        if self.caption_override is not None:
-            caption = self.caption_override
-        elif sample.get("json_path"):
-            caption = self._process_caption_json(sample["json_path"])
-        
-        if caption is None and sample.get("txt_path"):
-            caption = sample["txt_path"].read_text(encoding="utf-8").strip()
-            caption = self._process_caption_txt(caption)
-        
-        if caption is None:
-            caption = ""
-
-        # ARB 分桶
-        if self.bucket_mgr:
-            tw, th = self.bucket_mgr.get_bucket(img.width, img.height)
-        else:
-            tw = th = self.resolution
-
-        # 缩放裁剪
-        scale = max(tw / img.width, th / img.height)
-        nw, nh = int(img.width * scale), int(img.height * scale)
-        img = img.resize((nw, nh), Image.LANCZOS)
-
-        left = (nw - tw) // 2
-        top = (nh - th) // 2
-        img = img.crop((left, top, left + tw, top + th))
-
-        # 水平翻转增强
-        if self.flip_augment and random.random() > 0.5:
-            img = img.transpose(Image.FLIP_LEFT_RIGHT)
-
-        # 转 tensor [-1, 1]
-        arr = np.array(img).astype(np.float32) / 127.5 - 1.0
-        tensor = torch.from_numpy(arr).permute(2, 0, 1)
-
-        return {"pixel_values": tensor, "caption": caption}
-
-
-class RepeatDataset(Dataset):
-    """Kohya 风格数据集重复"""
-    def __init__(self, dataset, repeats=1):
-        self.dataset = dataset
-        self.repeats = max(1, int(repeats))
-
-    def __len__(self):
-        return len(self.dataset) * self.repeats
-
-    def __getitem__(self, idx):
-        return self.dataset[idx % len(self.dataset)]
-
-
-class MergedDataset(Dataset):
-    """合并主数据集与正则数据集（Kohya 风格 reg）"""
-    def __init__(self, main_dataset, reg_dataset, reg_weight: float = 1.0):
-        self.main_dataset = main_dataset
-        self.reg_dataset = reg_dataset
-        self.reg_weight = float(reg_weight)
-        self._main_len = len(main_dataset)
-        self._reg_len = len(reg_dataset)
-
-        # 为 BucketBatchSampler 构建 bucket_for_index
-        self.bucket_for_index = self._build_bucket_for_index()
-
-    def _get_cached_dataset(self, d):
-        if hasattr(d, "bucket_for_index"):
-            return d
-        if hasattr(d, "dataset"):
-            return self._get_cached_dataset(d.dataset)
-        return None
-
-    def _build_bucket_for_index(self):
-        main_cached = self._get_cached_dataset(self.main_dataset)
-        reg_cached = self._get_cached_dataset(self.reg_dataset)
-        buckets = []
-        if main_cached and main_cached.bucket_for_index:
-            main_base_len = len(main_cached.bucket_for_index)
-            for idx in range(self._main_len):
-                b = main_cached.bucket_for_index[idx % main_base_len]
-                buckets.append(b if b is not None else (0, 0))
-        else:
-            buckets.extend([(0, 0)] * self._main_len)
-        if reg_cached and reg_cached.bucket_for_index:
-            reg_base_len = len(reg_cached.bucket_for_index)
-            for idx in range(self._reg_len):
-                b = reg_cached.bucket_for_index[idx % reg_base_len]
-                buckets.append(b if b is not None else (0, 0))
-        else:
-            buckets.extend([(0, 0)] * self._reg_len)
-        return buckets
-
-    def __len__(self):
-        return self._main_len + self._reg_len
-
-    def __getitem__(self, idx):
-        if idx < self._main_len:
-            item = self.main_dataset[idx]
-            item["loss_weight"] = 1.0
-            return item
-        item = self.reg_dataset[idx - self._main_len]
-        item["loss_weight"] = self.reg_weight
-        return item
-
-
-class BucketBatchSampler:
-    """Batch sampler that groups samples by bucket so latents in each batch have the same size."""
-    def __init__(self, dataset, batch_size, drop_last=True, shuffle=True, seed=42):
-        self.dataset = dataset
-        self.batch_size = int(batch_size)
-        self.drop_last = bool(drop_last)
-        self.shuffle = bool(shuffle)
-        self.seed = int(seed)
-        self.epoch = 0
-        self._cached_dataset = self._get_cached_dataset(dataset)
-        self._base_len = len(self._cached_dataset) if self._cached_dataset else 0
-
-    def _get_cached_dataset(self, d):
-        if hasattr(d, "bucket_for_index"):
-            return d
-        if hasattr(d, "dataset"):
-            return self._get_cached_dataset(d.dataset)
-        return None
-
-    def set_epoch(self, epoch):
-        self.epoch = int(epoch)
-
-    def __len__(self):
-        n = len(self.dataset)
-        if self.drop_last:
-            return n // self.batch_size
-        return (n + self.batch_size - 1) // self.batch_size
-
-    def __iter__(self):
-        rng = random.Random(self.seed + self.epoch)
-        if self._cached_dataset is None:
-            indices = list(range(len(self.dataset)))
-            if self.shuffle:
-                rng.shuffle(indices)
-            for i in range(0, len(indices), self.batch_size):
-                batch = indices[i:i + self.batch_size]
-                if len(batch) < self.batch_size and self.drop_last:
-                    continue
-                yield batch
-            return
-
-        bucket_to_indices = {}
-        for idx in range(len(self.dataset)):
-            base_idx = idx % self._base_len
-            bucket = self._cached_dataset.bucket_for_index[base_idx]
-            if bucket is None:
-                bucket = (0, 0)
-            bucket_to_indices.setdefault(bucket, []).append(idx)
-
-        buckets = list(bucket_to_indices.keys())
-        if self.shuffle:
-            rng.shuffle(buckets)
-        for bucket in buckets:
-            indices = bucket_to_indices[bucket]
-            if self.shuffle:
-                rng.shuffle(indices)
-            for i in range(0, len(indices), self.batch_size):
-                batch = indices[i:i + self.batch_size]
-                if len(batch) < self.batch_size and self.drop_last:
-                    continue
-                yield batch
-
-
-class CachedLatentDataset(Dataset):
-    """Kohya 风格 npz 文件缓存的数据集"""
-    def __init__(self, base_dataset, vae, device, dtype, cache_dir=None):
-        import numpy as np
-        self.base_dataset = base_dataset
-        self.np = np
-        # 获取原始数据集的 samples 列表
-        self.samples = self._get_base_samples(base_dataset)
-        self.cache_dir = Path(cache_dir) if cache_dir else None
-        self.bucket_for_index = []
-        self._build_cache(vae, device, dtype)
-
-    def _get_base_samples(self, dataset):
-        """获取原始 ImageDataset 的 samples"""
-        if hasattr(dataset, "samples"):
-            return dataset.samples
-        elif hasattr(dataset, "dataset"):
-            return self._get_base_samples(dataset.dataset)
-        return []
-
-    def _get_npz_path(self, img_path):
-        """获取图像对应的 npz 缓存路径"""
-        img_path = Path(img_path)
-        return img_path.with_suffix(".npz")
-
-    def _is_cache_valid(self, img_path, npz_path):
-        """检查缓存是否有效（图像未修改，且格式含 latent 键）。
-        若为其他模型的不兼容缓存，则删除并返回 False。"""
-        if not npz_path.exists():
-            return False
-        if npz_path.stat().st_mtime < img_path.stat().st_mtime:
-            return False
-        try:
-            data = self.np.load(npz_path)
-            if "latent" not in data.files:
-                npz_path.unlink()
-                logger.debug(f"已删除不兼容缓存: {npz_path.name}")
-                return False
-        except Exception:
-            try:
-                npz_path.unlink()
-            except Exception:
-                pass
-            return False
-        return True
-
-    def _build_cache(self, vae, device, dtype):
-        """构建/加载 npz 缓存"""
-        logger.info("检查 VAE latent 缓存...")
-        to_encode = []
-        for i, sample in enumerate(self.samples):
-            img_path = sample["image"]
-            npz_path = self._get_npz_path(img_path)
-            if not self._is_cache_valid(img_path, npz_path):
-                to_encode.append(i)
-
-        if to_encode:
-            logger.info(f"需要编码 {len(to_encode)}/{len(self.samples)} 张图像...")
-            self._encode_and_save(to_encode, vae, device, dtype)
-        else:
-            logger.info(f"所有 {len(self.samples)} 张图像已缓存")
-
-        self._fill_bucket_for_index()
-
-    def _fill_bucket_for_index(self):
-        """Fill bucket_for_index for all samples (needed for BucketBatchSampler).
-        Uses latent spatial shape (h, w) as grouping key so batches have consistent tensor sizes."""
-        self.bucket_for_index = [None] * len(self.samples)
-        for i in range(len(self.samples)):
-            npz_path = self._get_npz_path(self.samples[i]["image"])
-            if not npz_path.exists():
-                continue
-            data = self.np.load(npz_path)
-            latent = data["latent"]
-            s = latent.shape
-            if len(s) == 5:
-                _, _, _, h, w = s
-            else:
-                _, _, h, w = s
-            self.bucket_for_index[i] = (int(h), int(w))
-
-    def _encode_and_save(self, indices, vae, device, dtype):
-        """编码图像并保存为 npz"""
-        for count, i in enumerate(indices):
-            item = self.base_dataset[i]
-            pixels = item["pixel_values"].unsqueeze(0).to(device, dtype=dtype)
-            _, _, ph, pw = pixels.shape
-            bucket_w, bucket_h = pw, ph
-            with torch.no_grad():
-                pixels_5d = pixels.unsqueeze(2)
-                latent = vae.model.encode(pixels_5d, vae.scale)
-            latent_np = latent.squeeze(0).cpu().float().numpy()
-            npz_path = self._get_npz_path(self.samples[i]["image"])
-            self.np.savez(npz_path, latent=latent_np, bucket_w=bucket_w, bucket_h=bucket_h)
-            if (count + 1) % 10 == 0 or count == len(indices) - 1:
-                logger.info(f"  编码进度: {count + 1}/{len(indices)}")
-
-    def __len__(self):
-        return len(self.samples)
-
-    def __getitem__(self, idx):
-        sample = self.samples[idx]
-        npz_path = self._get_npz_path(sample["image"])
-        data = self.np.load(npz_path)
-        latent = torch.from_numpy(data["latent"])
-        
-        # 获取 base_dataset 的引用（处理可能的嵌套）
-        base = self.base_dataset
-        while hasattr(base, "dataset"):
-            base = base.dataset
-        
-        # 处理 caption（正则集 caption_override 优先）
-        caption = None
-        if getattr(base, "caption_override", None) is not None:
-            caption = base.caption_override
-        elif sample.get("json_path") and hasattr(base, "_process_caption_json"):
-            caption = base._process_caption_json(sample["json_path"])
-        
-        if caption is None and sample.get("txt_path"):
-            caption = sample["txt_path"].read_text(encoding="utf-8").strip()
-            if hasattr(base, "_process_caption_txt"):
-                caption = base._process_caption_txt(caption)
-        
-        if caption is None:
-            caption = ""
-        
-        return {"latent": latent, "caption": caption}
-
-
-# ============================================================================
-# 训练时推理
-# ============================================================================
-
-@torch.no_grad()
-def sample_image(
-    model, vae, qwen_model, qwen_tokenizer, t5_tokenizer,
-    prompt, height=1024, width=1024, steps=25, cfg_scale=4.0,
-    negative_prompt=None,
-    sampler_name: str = "er_sde",
-    scheduler: str = "simple",
-    device="cuda",
-    dtype=torch.bfloat16,
-    step_callback=None,
-):
-    """训练时采样预览（尽量对齐 ComfyUI KSampler）
-    
-    Args:
-        negative_prompt: 负面提示词，默认使用标准负面提示词
-        sampler_name: 采样器（推荐：er_sde）
-        scheduler: 调度器（推荐：simple）
-    """
-    import numpy as np
-    from PIL import Image
-    model.eval()
-    
-    logger.info(f"[Debug] Sampling start. Prompt: {prompt[:50]}...")
-    
-    # Check VAE scale
-    if isinstance(vae.scale, list) and len(vae.scale) == 2:
-        m, s = vae.scale
-        logger.info(f"[Debug] VAE scale: mean_shape={m.shape}, std_inv_shape={s.shape}")
-        logger.info(f"[Debug] VAE scale values: mean={m.mean().item():.4f}, std_inv={s.mean().item():.4f}")
-
-    # 默认负面提示词 (参考 Anima Prompt Guide)
-    if negative_prompt is None:
-        negative_prompt = "worst quality, low quality, score_1, score_2, score_3, blurry, jpeg artifacts, bad anatomy, bad hands, bad feet, missing fingers, extra fingers, text, watermark, logo, signature, username, artist name, copyright name"
-
-    # 文本编码
-    try:
-        # 有条件 (positive prompt)
-        qwen_text = _build_qwen_text_from_prompt(prompt)
-        qwen_embeds, qwen_attn = encode_qwen(qwen_model, qwen_tokenizer, [qwen_text], device)
-        logger.info(f"[Debug] Qwen embeds: {qwen_embeds.shape}, mean={qwen_embeds.mean().item():.4f}")
-        
-        t5_ids, t5_attn, t5_w = tokenize_t5_weighted(t5_tokenizer, [prompt], max_length=512)
-        t5_ids = t5_ids.to(device)
-        t5_attn = t5_attn.to(device)
-        t5_w = t5_w.to(device, dtype=torch.float32)
-        cross_cond = model.preprocess_text_embeds(qwen_embeds, t5_ids)
-        if cross_cond.shape[1] < 512:
-            cross_cond = F.pad(cross_cond, (0, 0, 0, 512 - cross_cond.shape[1]))
-
-        # 无条件/负面提示词 (negative prompt)
-        qwen_text_uncond = _build_qwen_text_from_prompt(negative_prompt)
-        qwen_embeds_uncond, qwen_attn_uncond = encode_qwen(qwen_model, qwen_tokenizer, [qwen_text_uncond], device)
-        t5_ids_uncond, t5_attn_uncond, t5_w_uncond = tokenize_t5_weighted(t5_tokenizer, [negative_prompt], max_length=512)
-        t5_ids_uncond = t5_ids_uncond.to(device)
-        t5_attn_uncond = t5_attn_uncond.to(device)
-        t5_w_uncond = t5_w_uncond.to(device, dtype=torch.float32)
-        cross_uncond = model.preprocess_text_embeds(qwen_embeds_uncond, t5_ids_uncond)
-        if cross_uncond.shape[1] < 512:
-            cross_uncond = F.pad(cross_uncond, (0, 0, 0, 512 - cross_uncond.shape[1]))
-            
-    except Exception as e:
-        logger.error(f"[Debug] Encoding failed: {e}")
-        raise e
-
-    # sigmas（对齐 ComfyUI supported_models.Anima: shift=3.0, multiplier=1.0）
-    lat_h, lat_w = height // 8, width // 8
-    if str(scheduler).lower() != "simple":
-        logger.warning(f"采样 scheduler={scheduler} 未实现，回退 simple")
-    sigmas = _flow_sigmas_simple(steps, shift=3.0, device=device)
-
-    # 初始化噪声（ComfyUI CONST.noise_scaling: x = sigma*noise + (1-sigma)*latent_image；txt2img latent_image=0）
-    x = torch.randn(1, 16, 1, lat_h, lat_w, device=device, dtype=torch.float32) * float(sigmas[0])
-    logger.info(f"[Debug] Latents init: {x.shape}, mean={x.mean().item():.4f}, std={x.std().item():.4f}")
-
-    pad_mask = torch.zeros(1, 1, lat_h, lat_w, device=device, dtype=dtype)
-    device_type = "cuda" if str(device).startswith("cuda") else "cpu"
-
-    def denoise_fn(x_in: torch.Tensor, sigma_in: torch.Tensor) -> torch.Tensor:
-        if not torch.is_tensor(sigma_in):
-            sigma_in = torch.tensor(float(sigma_in), device=x_in.device, dtype=torch.float32)
-        sigma_b = sigma_in.view(1, 1).to(device=x_in.device, dtype=dtype)
-        sigma_5d = sigma_in.view(1, 1, 1, 1, 1).to(device=x_in.device, dtype=torch.float32)
-
-        with torch.autocast(device_type=device_type, dtype=dtype):
-            v_cond = model(x_in.to(device=x_in.device, dtype=dtype), sigma_b, cross_cond, padding_mask=pad_mask)
-            v_uncond = model(x_in.to(device=x_in.device, dtype=dtype), sigma_b, cross_uncond, padding_mask=pad_mask)
-            v = v_uncond + cfg_scale * (v_cond - v_uncond)
-
-        if torch.isnan(v).any():
-            raise RuntimeError("v contains NaN during sampling")
-
-        # CONST(flow): denoised x0 = x - sigma * v
-        return x_in - sigma_5d * v.float()
-
-    sampler_name_l = str(sampler_name).lower().strip()
-    logger.info(f"[Debug] Sampler={sampler_name_l}, Scheduler=simple, steps={steps}, cfg={cfg_scale}")
-
-    if sampler_name_l == "er_sde":
-        x = _sample_er_sde_const_x0(
-            denoise_fn, x, sigmas,
-            seed=None, s_noise=1.0, max_stage=3,
-            step_callback=step_callback,
-        )
-    else:
-        # fallback: 简化 Euler ODE（deterministic），与 flow 兼容
-        total = len(sigmas) - 1
-        for i in range(total):
-            sigma = float(sigmas[i])
-            sigma_next = float(sigmas[i + 1])
-            denoised = denoise_fn(x, sigmas[i])
-            if step_callback is not None:
-                try:
-                    step_callback(i, total, denoised)
-                except Exception:
-                    pass
-            d = (x - denoised) / max(sigma, 1e-6)
-            x = x + d * (sigma_next - sigma)
-
-    # VAE 解码
-    latents = x.to(device=device, dtype=dtype)
-    logger.info(f"[Debug] Final latents: mean={latents.mean().item():.4f}, std={latents.std().item():.4f}")
-    try:
-        images = vae.model.decode(latents, vae.scale)
-        images = images.squeeze(2)  # [B,C,H,W]
-        images = (images.clamp(-1, 1) + 1) / 2
-
-        # 转 PIL
-        img = images[0].permute(1, 2, 0).cpu().float().numpy()
-        img = (img * 255).clip(0, 255).astype(np.uint8)
-
-        model.train()
-        return Image.fromarray(img)
-    except Exception as e:
-        logger.error(f"[Debug] VAE decode failed: {e}")
-        raise e
-
-
-# ============================================================================
-# 训练辅助
-# ============================================================================
-
-def sample_t(bs, device, mode: str = "logit_normal", shift: float = 3.0) -> torch.Tensor:
-    """采样 Flow Matching 时间步 t ∈ (0, 1)。
-
-    mode:
-      logit_normal      — SD3/Anima 默认，偏向中间 t；shift>1 推向高噪声端
-      uniform           — 均匀采样，对细节端和结构端覆盖更均衡
-      logit_normal_low  — logit-normal 反向 shift，偏向低噪声/细节端
-      mode              — SD3 mode-distribution，集中在某个 sigma 附近
-    """
-    mode = (mode or "logit_normal").lower()
-    u = torch.sigmoid(torch.randn(bs, device=device))
-
-    if mode == "uniform":
-        return torch.rand(bs, device=device).clamp(1e-4, 1 - 1e-4)
-
-    if mode == "logit_normal_low":
-        s = max(float(shift), 1e-4)
-        u = (u * (1.0 / s)) / (1 + (1.0 / s - 1) * u)
-        return u.clamp(1e-4, 1 - 1e-4)
-
-    if mode == "mode":
-        s = float(shift)
-        u = 1 - u - s * (torch.cos(torch.pi * 0.5 * u) ** 2 - 1 + u)
-        return u.clamp(1e-4, 1 - 1e-4)
-
-    # logit_normal（默认）+ shift
-    s = float(shift)
-    u = (u * s) / (1 + (s - 1) * u)
-    return u.clamp(1e-4, 1 - 1e-4)
-
-
-def make_noise(
-    latents: torch.Tensor,
-    noise_offset: float = 0.0,
-    pyramid_iters: int = 0,
-    pyramid_discount: float = 0.35,
-) -> torch.Tensor:
-    """生成训练噪声，可叠加低频扰动。
-
-    noise_offset   — 给每样本/通道加低频偏移，缓解亮度均值偏差（SDXL 思路）
-    pyramid_iters  — 叠加多尺度低频噪声，帮助模型快速学习全局光照/构图；
-                     bilinear 插值避免 nearest 的块状结构干扰
-    """
-    noise = torch.randn_like(latents)
-
-    if noise_offset > 0:
-        shape = list(latents.shape)
-        for ax in range(2, latents.ndim):
-            shape[ax] = 1
-        offset = torch.randn(*shape, device=latents.device, dtype=latents.dtype)
-        noise = noise + noise_offset * offset
-
-    if pyramid_iters > 0:
-        try:
-            spatial = list(latents.shape[-2:])
-            cur = noise.clone()
-            for i in range(pyramid_iters):
-                r = 2 ** (i + 1)
-                sh, sw = max(spatial[0] // r, 1), max(spatial[1] // r, 1)
-                if latents.ndim == 5:
-                    extra = torch.randn(
-                        latents.shape[0], latents.shape[1], latents.shape[2], sh, sw,
-                        device=latents.device, dtype=latents.dtype,
-                    )
-                    extra = F.interpolate(
-                        extra.flatten(0, 1), size=spatial, mode="bilinear", align_corners=False,
-                    ).view(latents.shape[0], latents.shape[1], latents.shape[2], *spatial)
-                else:
-                    extra = torch.randn(latents.shape[0], latents.shape[1], sh, sw,
-                                        device=latents.device, dtype=latents.dtype)
-                    extra = F.interpolate(extra, size=spatial, mode="bilinear", align_corners=False)
-                cur = cur + extra * (pyramid_discount ** (i + 1))
-                if min(sh, sw) <= 1:
-                    break
-            noise = cur / cur.std().clamp(min=1e-6)
-        except Exception as exc:
-            logger.warning(f"pyramid_noise 失败，回退标准噪声: {exc}")
-
-    return noise
-
-
-def compute_loss_weight(
-    t: torch.Tensor,
-    scheme: str = "none",
-    min_snr_gamma: float = 5.0,
-    weight_cap_ratio: float = 0.0,
-) -> torch.Tensor:
-    """返回每样本 loss 权重 (B,)，Flow Matching CONST 调度下：SNR(t) = ((1-t)/t)^2。
-
-    scheme:
-      none          — 全 1，与原始行为一致
-      min_snr       — w = min(gamma/SNR, 1)，下调高 SNR 简单步（推荐基础款）
-      detail_inv_t  — w = 1/t clamp [1,5]，温和细节强化，小 batch + Prodigy 友好
-      cosmap        — SD3 cosmap weighting，中间 t 更均匀（max/min ≈ 1.81×）
-
-    weight_cap_ratio — batch 内 max/min 比上限（0=禁用），防单样本主导破坏 Prodigy d 估计
-    """
-    scheme = (scheme or "none").lower()
-    if scheme == "none":
-        return torch.ones_like(t)
-
-    eps = 1e-4
-    t_c = t.clamp(eps, 1 - eps)
-
-    if scheme == "min_snr":
-        snr = ((1 - t_c) / t_c) ** 2
-        w = torch.minimum(torch.tensor(float(min_snr_gamma), device=t.device) / snr, torch.ones_like(t_c))
-    elif scheme == "detail_inv_t":
-        w = (1.0 / t_c).clamp(min=1.0, max=5.0)
-    elif scheme == "cosmap":
-        bot = (1 - 2 * t_c + 2 * t_c ** 2).clamp(min=eps)
-        w = 2.0 / (math.pi * bot)
-    else:
-        return torch.ones_like(t)
-
-    if weight_cap_ratio and weight_cap_ratio > 1.0:
-        w_min = w.min().clamp(min=eps)
-        w = w.clamp(max=w_min * float(weight_cap_ratio))
-
-    return w
-
-
-def collate_fn(batch):
-    """DataLoader collate"""
-    pixels = torch.stack([b["pixel_values"] for b in batch])
-    captions = [b["caption"] for b in batch]
-    result = {"pixel_values": pixels, "captions": captions}
-    if "loss_weight" in batch[0]:
-        result["loss_weight"] = torch.tensor([b["loss_weight"] for b in batch], dtype=torch.float32)
-    return result
-
-
-def collate_fn_cached(batch):
-    """DataLoader collate for cached latents"""
-    latents = torch.stack([b["latent"] for b in batch])
-    captions = [b["caption"] for b in batch]
-    result = {"latents": latents, "captions": captions}
-    if "loss_weight" in batch[0]:
-        result["loss_weight"] = torch.tensor([b["loss_weight"] for b in batch], dtype=torch.float32)
-    return result
-
-
-# ============================================================================
-# 参数解析
-# ============================================================================
-
-def parse_args():
-    """从 studio.schema.TrainingConfig 自动生成 parser；额外补 schema 之外的
-    CLI-only 开关（auto-install / interactive / no-live-curve / 已弃用的
-    --repeats 和 --reg-repeats）。
-    """
-    from studio.argparse_bridge import build_parser
-    from studio.schema import TrainingConfig
-
-    p = build_parser(TrainingConfig, prog="anima_train", description="Anima LoRA Trainer v2")
-    # schema 之外的 CLI-only 开关
-    p.add_argument("--auto-install", action="store_true", help="自动安装缺失依赖")
-    p.add_argument("--interactive", action="store_true", help="交互模式，提示输入缺失参数")
-    p.add_argument("--no-live-curve", action="store_true", help="禁用实时 Loss 曲线刷新")
-    # PP6.1 — 监控状态文件路径；不传则默认写到 output_dir/monitor_state.json
-    # 注：--no-monitor / --monitor-host / --monitor-port / --no-browser 由 schema
-    # 自动从 TrainingConfig 字段生成（保留只为兼容旧 yaml，运行时忽略）。
-    p.add_argument(
-        "--monitor-state-file",
-        type=str,
-        default=None,
-        help="训练监控 state.json 输出路径（默认 output_dir/monitor_state.json）",
-    )
-    # 已弃用：每图重复改用文件夹名前缀（如 5_concept）
-    p.add_argument("--repeats", type=int, default=1, help=argparse.SUPPRESS)
-    p.add_argument("--reg-repeats", type=int, default=1, help=argparse.SUPPRESS)
-    return p.parse_args()
-
-
-# ============================================================================
-# 交互模式辅助函数
-# ============================================================================
-
-def _try_rich():
-    try:
-        from rich.prompt import Prompt, Confirm
-        return Prompt, Confirm
-    except Exception:
-        return None, None
-
-
-def _ask_str(label, default=""):
-    Prompt, _ = _try_rich()
-    if Prompt:
-        return Prompt.ask(label, default=default) if default else Prompt.ask(label)
-    raw = input(f"{label}{f' [{default}]' if default else ''}: ").strip()
-    return raw or default
-
-
-def _ask_bool(label, default=False):
-    _, Confirm = _try_rich()
-    if Confirm:
-        return Confirm.ask(label, default=default)
-    raw = input(f"{label} [{'Y/n' if default else 'y/N'}]: ").strip().lower()
-    if not raw:
-        return default
-    return raw in ("y", "yes", "1", "true", "t")
-
-
-def _ask_int(label, default):
-    while True:
-        raw = _ask_str(label, str(default))
-        try:
-            return int(raw)
-        except ValueError:
-            print("Please enter an integer.")
-
-
-def _ask_float(label, default):
-    while True:
-        raw = _ask_str(label, str(default))
-        try:
-            return float(raw)
-        except ValueError:
-            print("Please enter a number.")
-
-
-def _guess_default_paths():
-    """猜默认模型路径（仅在用户没在 yaml/CLI 显式指定时用）。
-
-    根目录：优先 `secrets.models.root`（Studio 设置页配置），否则 `REPO_ROOT/models/`
-    （与 schema.py 默认 + WD14 已用的 `models/wd14/` 对齐）。
-
-    Transformer：用户可能装多个 Anima 版本（preview / preview2 / preview3-base），
-    按 ANIMA_VARIANTS 顺序找第一个存在的（latest 优先）。
-    """
-    repo_root = Path(__file__).resolve().parent
-    # secrets 不一定可 import（直接 CLI 跑训练时 studio package 可用；其他场景兜底）
-    base: Optional[Path] = None
-    transformer_path: str = ""
-    try:
-        from studio.services.model_downloader import find_anima_main, models_root
-        base = models_root()
-        existing = find_anima_main(base)
-        if existing:
-            transformer_path = str(existing)
-    except Exception:
-        base = repo_root / "models"
-    if not base:
-        base = repo_root / "models"
-    if not transformer_path:
-        # services 不可用 / 都没下载 → 给最新版默认名作为提示，方便用户填路径
-        candidate = base / "diffusion_models" / "anima-preview3-base.safetensors"
-        transformer_path = str(candidate) if candidate.exists() else ""
-
-    vae = base / "vae" / "qwen_image_vae.safetensors"
-    qwen = base / "text_encoders"
-    return {
-        "transformer": transformer_path,
-        "vae": str(vae) if vae.exists() else "",
-        "qwen": str(qwen) if qwen.exists() else "",
-    }
-
-
-def prompt_for_args(args):
-    """交互式提示输入缺失参数"""
-    defaults = _guess_default_paths()
-    args.data_dir = args.data_dir or _ask_str("数据集目录 (images + .txt)", "")
-    args.transformer_path = args.transformer_path or _ask_str("Transformer 路径 (.safetensors)", defaults["transformer"])
-    args.vae_path = args.vae_path or _ask_str("VAE 路径 (.safetensors)", defaults["vae"])
-    args.text_encoder_path = args.text_encoder_path or _ask_str("Qwen 模型目录", defaults["qwen"])
-    args.output_dir = _ask_str("输出目录", args.output_dir)
-    args.output_name = _ask_str("输出名称", args.output_name)
-    args.resolution = _ask_int("分辨率", args.resolution)
-    args.batch_size = _ask_int("Batch size", args.batch_size)
-    args.grad_accum = _ask_int("梯度累积", args.grad_accum)
-    args.learning_rate = _ask_float("学习率", args.learning_rate)
-    args.grad_checkpoint = _ask_bool("启用梯度检查点?", args.grad_checkpoint)
-    args.epochs = _ask_int("Epochs", args.epochs)
-    args.max_steps = _ask_int("最大步数 (0=无限制)", args.max_steps)
-    args.lora_rank = _ask_int("LoRA rank", args.lora_rank)
-    args.lora_alpha = _ask_float("LoRA alpha", args.lora_alpha)
-    args.loss_curve_steps = _ask_int("Loss 曲线步数 (0=禁用)", args.loss_curve_steps)
-    args.auto_install = _ask_bool("自动安装缺失依赖?", args.auto_install)
-    args.save_every_epoch = _ask_bool("每个 epoch 保存?", args.save_every_epoch)
-    args.mixed_precision = _ask_str("混合精度 (bf16/fp32)", args.mixed_precision)
-    return args
+# ─── Re-exports for sister script / tests (ADR 0003 PR-A) ────────────────────
+# 这些名字被 anima_daemon / anima_generate / anima_reg_ai (`import anima_train as _T`
+# 然后 _T.X) 以及 tests/test_anima_train_migration.py 等直接读取。新代码请
+# 直接 import 子模块，不要再依赖 anima_train 顶层。
+from training.bootstrap import (  # noqa: E402
+    apply_yaml_config,
+    ensure_dependencies,
+    init_progress,
+    load_yaml_config,
+)
+from training.observability import (  # noqa: E402
+    WandBMonitor,
+    init_wandb_monitor,
+    render_curve_panel,
+    render_loss_curve,
+)
+from training.model_loading import (  # noqa: E402
+    _load_safetensors_state_dict,
+    _load_weights_best_effort,
+    _pick_best_prefix_remap,
+    _strip_prefixes,
+    enable_xformers,
+    find_diffusion_pipe_root,
+    forward_with_optional_checkpoint,
+    load_module_from_path,
+    resolve_path_best_effort,
+)
+from training.text_encoding import (  # noqa: E402
+    _build_qwen_text_from_prompt,
+    _parse_weighted_tag,
+    encode_qwen,
+    tokenize_t5_weighted,
+)
+from training.state import load_training_state, save_training_state  # noqa: E402
+from training.models import (  # noqa: E402
+    ensure_models_namespace,
+    load_anima_model,
+    load_text_encoders,
+    load_vae,
+)
+from training.sampling import sample_image  # noqa: E402
+from training.dataset import (  # noqa: E402
+    BucketBatchSampler,
+    BucketManager,
+    CachedLatentDataset,
+    ImageDataset,
+    MergedDataset,
+    RepeatDataset,
+    collate_fn,
+    collate_fn_cached,
+)
+from training.cli import (  # noqa: E402
+    parse_args,
+    prompt_for_args,
+)
+from training.timestep_sampling import sample_t  # noqa: E402
+from training.noise import make_noise  # noqa: E402
+from training.loss_weighting import compute_loss_weight  # noqa: E402
 
 
 # ============================================================================
